pyproject.toml
[project]
name = "mcp-use-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-groq>=0.3.2",
    "mcp-use>=1.2.13",
    "python-dotenv>=1.1.0",
]
